requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.7.2
attrs==23.1.0
black==23.9.1
blessed==1.20.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Django==4.2.5
django-debug-toolbar==4.2.0
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
idna==3.4
inflection==0.5.1
jinxed==1.2.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==3.5.3
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
six==1.16.0
sqlparse==0.4.4
stripe==6.7.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
wcwidth==0.2.8
